#199 add deps to useEffect, rename ServiceCard.js to index.js
